requirements.txt
speechrecognition
pyaudio
pyperclip
pyautogui
pystray
Pillow